+++ dummy/any.ts
@@ -2,24 +2,14 @@
 // This module is browser compatible.
 
 import { equality } from "../helper/equal.ts";
-import { isObject, isSymbol } from "../deps.ts";
+import { isObject } from "../deps.ts";
 import type { Equality } from "../helper/equal.ts";
 import { isAnyNumber } from "./any_number.ts";
 import { isAnyString } from "./any_string.ts";
 import { isAnyBoolean } from "./any_boolean.ts";
 import { isAnyFunction } from "./any_function.ts";
-
-function isBigInt(value: unknown): value is bigint {
-  return typeof value === "bigint";
-}
-
-function isAnyBigInt(value: unknown): value is bigint | BigInt {
-  return isBigInt(value) || value instanceof BigInt;
-}
-
-function isAnySymbol(value: unknown): value is symbol | Symbol {
-  return isSymbol(value) || value instanceof Symbol;
-}
+import { isAnyBigInt } from "./any_big_int.ts";
+import { isAnySymbol } from "./any_symbol.ts";
 
 function equalAnyOf(a: unknown, b: Function): boolean {
   switch (b) {
@@ -57,6 +47,7 @@ class Any implements Equality {
     return equalAnyOf(actual, this.object);
   }
 
+  /** display name */
   toString(): string {
     return "Any";
   }
@@ -75,5 +66,4 @@ export {
   isAnyFunction,
   isAnyString,
   isAnySymbol,
-  isBigInt,
 };