+++ dummy/any_test.ts
@@ -1,13 +1,5 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
-import {
-  Any,
-  any,
-  equalAnyOf,
-  isAnyBigInt,
-  isAnyString,
-  isAnySymbol,
-  isBigInt,
-} from "./any.ts";
+import { Any, any, equalAnyOf, isAnyString } from "./any.ts";
 import { assertEquals, assertExists } from "../dev_deps.ts";
 import { equality } from "../helper/equal.ts";
 
@@ -23,63 +15,6 @@ Deno.test("isAnyString", () => {
   ];
 
   table.forEach(([value, result]) => assertEquals(isAnyString(value), result));
-});
-
-Deno.test("isAnySymbol", () => {
-  const table: [
-    ...Parameters<typeof isAnySymbol>,
-    ReturnType<typeof isAnySymbol>,
-  ][] = [
-    [Symbol.for(""), true],
-    [Symbol("test"), true],
-    [Symbol, false],
-  ];
-
-  table.forEach(([value, result]) => assertEquals(isAnySymbol(value), result));
-});
-
-Deno.test("isAnyBigInt", () => {
-  const table: [
-    ...Parameters<typeof isAnyBigInt>,
-    ReturnType<typeof isAnyBigInt>,
-  ][] = [
-    [1n, true],
-    [BigInt(0n), true],
-    [BigInt, false],
-  ];
-
-  table.forEach(([value, result]) => assertEquals(isAnyBigInt(value), result));
-});
-
-Deno.test("isAnyBigInt", () => {
-  const table: [
-    ...Parameters<typeof isAnyBigInt>,
-    ReturnType<typeof isAnyBigInt>,
-  ][] = [
-    [1n, true],
-    [BigInt(0n), true],
-    [BigInt, false],
-    [1, false],
-  ];
-
-  table.forEach(([value, result]) => assertEquals(isAnyBigInt(value), result));
-});
-
-Deno.test("isBitInt", () => {
-  const table: [
-    ...Parameters<typeof isBigInt>,
-    ReturnType<typeof isBigInt>,
-  ][] = [
-    [1n, true],
-    [0n, true],
-    [0, false],
-    [1, false],
-    ["", false],
-    [{}, false],
-    [null, false],
-  ];
-
-  table.forEach(([value, result]) => assertEquals(isBigInt(value), result));
 });
 
 Deno.test("equalAnyOf", () => {